refactor: 判断存在 .vimrc 的情况下进行备份

+++ init.sh
@@ -1,2 +1,7 @@
-mv ~/.vimrc ~/.vimrc.bak
+#!/bin/bash
+
+if [ -f ~/.vimrc ]; then
+	mv ~/.vimrc ~/.vimrc.bak
+fi
+
 mv .vimrc ~/.vimrc
